Fix server crash due to referencing client classes 1. Moved gui construction to the client 2. Marked ImageUtils#getOrLoadImage as a client method (as this class is referenced across environments)
